src/main/java/com/miguel/library/Config/CustomDetailsService.java
package com.miguel.library.Config;

import com.miguel.library.Exceptions.ExceptionObjectNotFound;
import com.miguel.library.model.User;
import com.miguel.library.services.IUserService;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.context.annotation.Configuration;
import org.springframework.security.core.userdetails.UserDetails;
import org.springframework.security.core.userdetails.UserDetailsService;
import org.springframework.security.core.userdetails.UsernameNotFoundException;
import org.springframework.stereotype.Service;

import java.util.Objects;

@Service
public class CustomDetailsService implements UserDetailsService {

    @Autowired
    private IUserService userService;

    @Override
    public UserDetails loadUserByUsername(String email) throws UsernameNotFoundException {

        User user = userService.searchByEmail(email);

        if (Objects.isNull(user)) {
            throw new ExceptionObjectNotFound("User Not Found");
        }

        return new org.springframework.security.core.userdetails.User(
                user.getUsername(),
                user.getPassword(),
                user.getRole().getAuthorities()
        );
    }
}
